Give in to evil (add constructors)
- Change `Object *` to be a union
- Add `Object_Header` for querying of common information from an `Object
  *` that all subjects also hold
- Add `Value_Type` which manually extends the public-facing `lulu_Type`
- Remove `*_make()` functions, replace with constructors
- Replace `*_eq()` functions with appropriate operator overloads except
  for `slice_eq()` due to its non-trivial but templated nature

+++ src/compiler.hpp
@@ -9,7 +9,7 @@ struct Compiler {
     lulu_VM &vm;
     Parser  &parser; // All compilers share the same parser.
     Chunk   &chunk;  // Compilers do not own their chunks.
-    u8       free_reg;
+    u16      free_reg;
 };
 
 Compiler
@@ -57,7 +57,7 @@ compiler_add_constant(Compiler &c, OString *s);
  *      `MAX_REG`.
  */
 void
-compiler_reserve_reg(Compiler &c, u16 n);
+compiler_reserve_reg(Compiler &c, int n);
 
 
 /**
